cosmetic refactoring

+++ sm-core/src/main/java/com/salesmanager/core/business/modules/integration/shipping/impl/CustomShippingQuoteRules.java
@@ -94,7 +94,7 @@ public class CustomShippingQuoteRules implements ShippingQuoteModule {
 		for(PackageDetails pack : packages) {
 			weight = weight + pack.getShippingWeight();
 			Double tmpVolume = pack.getShippingHeight() * pack.getShippingLength() * pack.getShippingWidth();
-			if(volume == null || tmpVolume.doubleValue() > volume.doubleValue()) { //take the largest volume
+			if(volume == null || tmpVolume > volume) { //take the largest volume
 				volume = tmpVolume;
 			} 
 			//largest size
@@ -102,9 +102,9 @@ public class CustomShippingQuoteRules implements ShippingQuoteModule {
 			sizeList.add(pack.getShippingHeight());
 			sizeList.add(pack.getShippingWeight());
 			sizeList.add(pack.getShippingLength());
-			Double maxSize = (Double)Collections.max(sizeList);
-			if(size==null || maxSize.doubleValue() > size.doubleValue()) {
-				size = maxSize.doubleValue();
+			Double maxSize = Collections.max(sizeList);
+			if(size==null || maxSize > size) {
+				size = maxSize;
 			}
 		}
 		
@@ -121,7 +121,7 @@ public class CustomShippingQuoteRules implements ShippingQuoteModule {
 		}
 		
 		if(distance!=null) {
-			double ddistance = distance.doubleValue();
+			double ddistance = distance;
 			long ldistance = (long)ddistance;
 			inputParameters.setDistance(ldistance);
 		}
